+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Corrected normalization behavior for scene thumbnails and annotation projects [#5569](https://github.com/raster-foundry/raster-foundry/pull/5569)
 - Restored coordinate space to 4096x4096 in label MVT [#5569](https://github.com/raster-foundry/raster-foundry/pull/5569)
+- Fixed campaign STAC exports so that classes are added to label STAC item and label data [#5570](https://github.com/raster-foundry/raster-foundry/pull/5570)
 
 ## [1.62.0] - 2021-04-14
 ### Changed